requirements.txt
beautifulsoup4==4.12.0
bs4==0.0.1
PyYAML==6.0
requests==2.28.2

